DifferentialDiagnosis/DifferentialDiagnosis/requirements.txt
opencv-python==4.5.1.48
tensorflow==2.4.1
google-cloud-storage==1.37.1
